requirements.txt
streamlit
langchain
langchain-core
langchain-mistralai
python-dotenv
pydantic